Fix sdk tests
